[refactor] 충돌 해결

+++ src/main/java/dev/woori/wooriBank/domain/account/dto/response/UserAccountResDto.java
@@ -1,7 +1,5 @@
 package dev.woori.wooriBank.domain.account.dto.response;
 
-import java.math.BigDecimal;
-
 /**
  * 회원 생성 + 계좌 개설 응답 DTO
  */
@@ -12,6 +10,6 @@ public record UserAccountResDto(
         String phoneNumber,
         Long accountId,
         String accountNumber,
-        BigDecimal balance
+        Long balance
 ) {
 }
